mh2lm: commonize
The only difference between mh2lm and mh2lm_5g is... you guessed it,
5G.

This difference in hardware is configured by a Qualcomm blob, so there
is no point in having separate trees for the same device.

Also, a mh2lm_5g tree never existed, and mh2lm just works on both
anyway.

Change-Id: I07c611f7ac071eb149eb4c7abf93c0241e79df76